Rename: letterStatus 이동
